+++ packages/adherev-ui-searchtable/src/searchtableimplement.tsx
@@ -2,7 +2,11 @@ import Vue, { CreateElement, VNode } from 'vue';
 // @ts-ignore
 import { mapState, mapMutations, mapActions, cleanMixin } from '@ctsj/vuexgenerator';
 import { TableRowSelection } from 'ant-design-vue/lib/table/interface';
-import SearchTable, { NUMBER_GENERATOR_RULE_CONTINUITY } from './searchtable';
+import SearchTable, {
+  ROW_SELECTION_NORMAL_MODE,
+  ROW_SELECTION_CONTINUOUS_MODE,
+  NUMBER_GENERATOR_RULE_CONTINUITY,
+} from './searchtable';
 
 const selectorPrefix = 'adherev-ui-searchtableimplement';
 
@@ -23,6 +27,7 @@ export default (serviceName) =>
           ...this.getParams(),
         },
         selectedRowKeys: [],
+        selectedRows: [],
       };
     },
     computed: serviceName
@@ -127,6 +132,12 @@ export default (serviceName) =>
           return NUMBER_GENERATOR_RULE_CONTINUITY;
         },
         /**
+         * getNumberGeneratorRule - 获取符号列的生成规则
+         */
+        getRowSelectionMode(): Symbol {
+          return ROW_SELECTION_NORMAL_MODE;
+        },
+        /**
          * getTableNumberColumnWidth
          * @override
          * @description - 表格序号列的宽度
@@ -185,10 +196,46 @@ export default (serviceName) =>
          * @description - 获取表格行选择对象
          */
         getRowSelection(): TableRowSelection<object> {
+          const self = this;
+
+          function filter(this: any, selected: boolean, records: Array<any>): void {
+            const rowKey = self.getRowKey();
+
+            if (selected) {
+              // add
+              self.selectedRowKeys = [
+                ...self.selectedRowKeys,
+                ...records.map((r) => r[rowKey]),
+              ];
+              self.selectedRows = [...self.selectedRows, ...records];
+            } else {
+              // remove
+              self.selectedRows = self.selectedRows.filter(
+                (row) => !records.find((r) => r[rowKey] === row[rowKey]),
+              );
+              self.selectedRowKeys = self.selectedRowKeys.filter(
+                (key) => !records.find((r) => r[rowKey] === key),
+              );
+            }
+          }
+
           return {
             selectedRowKeys: this.selectedRowKeys,
-            onChange: (selectedRowKeys: Array<any>) => {
+            onChange: (selectedRowKeys: Array<any>, selectedRows: Array<any>) => {
+              if (this.getRowSelectionMode() === ROW_SELECTION_CONTINUOUS_MODE) return;
+
               this.selectedRowKeys = selectedRowKeys;
+              this.selectedRows = selectedRows;
+            },
+            onSelect: (record, selected) => {
+              if (this.getRowSelectionMode() === ROW_SELECTION_NORMAL_MODE) return;
+
+              filter(selected, [record]);
+            },
+            onSelectAll: (selected, selectedRows, changeRows) => {
+              if (this.getRowSelectionMode() === ROW_SELECTION_NORMAL_MODE) return;
+
+              filter(selected, changeRows);
             },
           };
         },